Change castTo/castFrom to not allocate
Currently castTo/castFrom compile down to IL methods that take a Teq and return an FSharpFunc, this always allocates. Removing these parens instead causes fsc to compile them as a method that takes a Teq and an 'a or 'b. Given we nearly never partially apply castTo/castFrom this helps with overall allocations.